Case 1, single student data file
This code goes through the JSON file of a single student and analyzes the student's actions when engaged in solving Case 1, challenges 2 and 3. In addition to actions related to entering and exiting these challenges, the code analyzes all gender changes and alleles changes, and scores them based on their utility in view of the current goal dragon the student is trying to match. The reports are displayed on a a primitive, barely functional web page and all scripts and data are locally stored.